move to `api::rest::schema`
